refactor: move hooks to own dir
