+++ app/(tabs)/siktemerker/components/CalculateMarksModal.tsx
@@ -3,8 +3,8 @@ import { FontAwesomeIcon } from '@fortawesome/react-native-fontawesome';
 import { Keyboard, Modal, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import { Button, Input } from '@/components/common';
 import { CalculatedMarks, MarksResult } from '@/types';
-import { useCalculateMarks, formatNumber, getLocalStorage, storeLocalStorage } from '@/utils';
-import { useCalcMarksForm } from './useCalcMarksForm';
+import { formatNumber, getLocalStorage, storeLocalStorage, useCalculateMarks } from '@/utils';
+import { useCalcMarksForm } from '../hooks/useCalcMarksForm';
 import { faRulerHorizontal } from '@fortawesome/free-solid-svg-icons/faRulerHorizontal';
 import { faCrosshairs } from '@fortawesome/free-solid-svg-icons/faCrosshairs';
 import { faChevronDown } from '@fortawesome/free-solid-svg-icons/faChevronDown';
