+++ app/(tabs)/siktemerker/components/index.ts
@@ -1,5 +1,3 @@
 export { default as ConfirmRemoveMarks } from './ConfirmRemoveMarks';
-export { default as MarksForm } from './MarksForm';
+export { default as MarksForm } from './marksForm/MarksForm';
 export { default as MarksTable } from './MarksTable';
-export { useCalcForm } from './useCalcForm';
-export { useCalcMarksForm } from './useCalcMarksForm';
