+++ app/(tabs)/siktemerker/hooks/useCalcForm.ts
