+++ app/(tabs)/siktemerker/hooks/useCalcMarksForm.ts
